services/crawler/requirements.txt
beautifulsoup4
TextBlob
feedparser
requests
pymongo>=4.0
python-dateutil
